Feature: recebe  as receitas e renderiza o respectivo componente passando a receita
