+++ src/App.js
@@ -11,6 +11,7 @@ import Comidas from './pages/Comidas';
 import Explorar from './pages/Explorar';
 import Bebidas from './pages/Bebidas';
 import Login from './pages/Login';
+import MealsProvider from './context/MealsProvider';
 
 function App() {
   return (
@@ -18,11 +19,13 @@ function App() {
       <BrowserRouter>
         <Switch>
           <Route exact path="/" component={ Login } />
-          <Route exact path="/comidas" component={ Comidas } />
+          <MealsProvider>
+            <Route exact path="/comidas" component={ Comidas } />
+            <Route exact path="/comidas/{id-da-receita}" />
+            <Route exact path="/comidas/{id-da-receita}/in-progress" />
+          </MealsProvider>
           <Route exact path="/bebidas" component={ Bebidas } />
-          <Route exact path="/comidas/{id-da-receita}" />
           <Route exact path="/bebidas/{id-da-receita}" />
-          <Route exact path="/comidas/{id-da-receita}/in-progress" />
           <Route exact path="/bebidas/{id-da-receita}/in-progress" />
           <Route exact path="/explorar" component={ Explorar } />
           <Route
