--- src/context/MealsProvider.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import MealsContext from './MealsContext';
-
-export default function MealsProvider() {
-  return (
-    <MealsContext.Provider>
-      { Children }
-    </MealsContext.Provider>
-  );
-}
