+++ src/pages/Comidas.jsx
@@ -1,29 +1,23 @@
-import React, { useState, useEffect } from 'react';
+import React, { useContext } from 'react';
 
 import Header from '../components/Header';
 import CategoryFilters from '../components/CategoryFilters';
-// import FoodCard from '../components/FoodCard';
+import FoodCard from '../components/FoodCard';
 import '../styles/Comidas.css';
+import MealsContext from '../context/MealsContext';
 
 function Comidas() {
-  const [mealsRecipes, setMealsRecipes] = useState([]);
-  useEffect(() => {
-    const fetchFoods = async () => {
-      const response = await fetch('https://www.themealdb.com/api/json/v1/1/search.php?s=');
-      const { meals } = await response.json();
-      setMealsRecipes(meals);
-    };
-    fetchFoods();
-  }, []);
+  const { mealsRecipes } = useContext(MealsContext);
 
-  console.log(mealsRecipes);
   return (
     <div className="comidasPageContainer">
       <Header title="Comidas" showSearchIcon />
       <CategoryFilters />
       <div className="foodRecipesContainer">
         {
-          mealsRecipes.map((recipe, index) => <h1 key={ index }>{ recipe.idMeal }</h1>)
+          mealsRecipes.map(
+            (recipe, index) => <FoodCard key={ index } recipe={ recipe } />,
+          )
         }
       </div>
     </div>
